#896 - runtime: make debug feature useful
