+++ tests/regression/namespaces/mu/core
@@ -36,4 +36,3 @@
 (mu:type-of mu:find-namespace)	:func
 (mu:type-of mu:intern)	:func
 (mu:type-of mu:make-namespace)	:func
-(mu:type-of mu:namespace-symbols)	:func
