+++ tools/crossref/crossref.l
@@ -15,7 +15,7 @@
                                   ,(core:type-of (mu:symbol-value symbol))
                                   ,(mu:symbol-value symbol)))
                    (core:%format stream "unbound~Tunbound~T~A~%" `(,(mu:symbol-name symbol)))))
-           (mu:namespace-symbols (mu:find-namespace "core")))
+           (mu/core:ns-symbols (mu:find-namespace "core")))
          (mu:close stream))
        (mu:open :file :output file-name :t))
       :t))
