+++ tools/metrics/symbols.l
@@ -11,7 +11,7 @@
            (:lambda (symbol)
              (core:%format stream "~A~%" `(,(mu:symbol-name symbol)))
              (core:%format stream "~A:~A~%" `(,ns ,(mu:symbol-name symbol))))
-           (mu:namespace-symbols (mu:find-namespace ns)))
+           (core:ns-symbols (mu:find-namespace ns)))
           (mu:close stream))
        (mu:open :file :output file-name :t))
       :t))
