+++ tools/reference/reference.l
@@ -16,7 +16,7 @@
                                    ,(mu:symbol-value symbol)
                                    ,(mu:repr symbol)))
                    (core:%format stream "unbound~Tunbound~T~A~Tunbound~%" `(,(mu:symbol-name symbol)))))
-           (mu:namespace-symbols (mu:find-namespace ns)))
+           (core:ns-symbols (mu:find-namespace ns)))
          (mu:close stream))
        (mu:open :file :output file-name :t))
       :t))
